add new logic to avoid address calculation
